dearpygui-spectrum-app/requirements.txt
DearPyGui
matplotlib
pyserial
psutil